Adding metadata editor modal
